tests for nested arrays
